upgrade exporter to v3
