--- ExporterAddon/constants/boundaries.js
@@ -1,6 +1,0 @@
-const boundaries = {
-    currentExtent: "currentExtent",
-    completeDataset: "completeDataset"
-};
-
-export default boundaries;
