--- ExporterAddon/constants/exportformats.js
@@ -1,8 +1,0 @@
-const exportformats = {
-    geoJson: "geojson",
-    gml: "gml",
-    shp: "shp",
-    gpkg: "gpkg"
-};
-
-export default exportformats;
