--- ExporterAddon/constants/geoPackageDataTypes.js
@@ -1,19 +1,0 @@
-const GeoPackageDataType = {
-    boolean: "BOOLEAN",
-    tinyint: "TINYINT",
-    smallint: "SMALLINT",
-    mediumint: "MEDIUMINT",
-    int: "INT",
-    integer: "INTEGER",
-    float: "FLOAT",
-    double: "DOUBLE",
-    real: "REAL",
-    text: "TEXT",
-    blob: "BLOB",
-    date: "DATE",
-    datetime: "DATETIME",
-    string: "TEXT",
-    number: "INT"
-};
-
-export default GeoPackageDataType;
