--- ExporterAddon/constants/layertypes.js
@@ -1,7 +1,0 @@
-const layertypes = {
-    wfs: "WFS",
-    draw: "DRAW",
-    geoJson: "GeoJSON"
-};
-
-export default layertypes;
