--- ExporterAddon/constants/steps.js
@@ -1,7 +1,0 @@
-const steps = {
-    selectLayer: "selectLayer",
-    downloadLayer: "downloadLayer",
-    selectExternalWfsLayer: "selectExternalWfsLayer"
-};
-
-export default steps;
