--- ExporterAddon/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "exporteraddon",
-  "version": "1.0.0",
-  "description": "A generic exporter addon for Masterportal",
-  "main": "index.js",
-  "directories": {
-    "doc": "doc",
-    "test": "tests"
-  },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@crmackey/shp-write": "^0.4.5"
-  }
-}
