+++ README.md
@@ -17,7 +17,8 @@ Please check the `README` or `doc` files within the respective addon folder.
 | fontawesome         | Add fontawesome as icon library                       | v2.5²       | [LGB](https://geobasis-bb.de/lgb/de/geodaten/portale-anwendungen/geoportal-brandenburg/#) |
 | simpleLineChart     | Renders data points on simple chart.js component      | v3.8.0       | [FOSSGIS WS](https://github.com/terrestris/masterportal-ws)  |
 | tourGuide     | This addon enables a customised tour through the masterportal. Elements can be highlighted and provided with descriptions and graphics. Technically, it is basically a wrapper for [intro.js](https://introjs.com/) with minor adaptions to fit the masterportal design.      | v3.8.0       | [Meckenheim Geoportal](https://geoportal.meckenheim.de)  |
-| Importer (v3)           | Import common geodata formats and OGC services        | v.3.13.0    |      |
+| importer (v3)       | Import common geodata formats and OGC services        | v.3.13.0    |      |
+| exporter (v3)       | Export WFS and vector layer to common geodata formats | v.3.13.0    |      |
 
 ¹ Should work on older versions that use the new Vue.js framework.  
 ² Last tested version. Should work on more recent versions that still use Backbone.js framework.
