+++ exporter/constants/exportformats.js
@@ -1,0 +1,8 @@
+const exportformats = {
+    geoJson: "geojson",
+    gml: "gml",
+    shp: "shp",
+    gpkg: "gpkg"
+};
+
+export default exportformats;
