+++ exporter/constants/layertypes.js
@@ -1,0 +1,8 @@
+const layertypes = {
+    wfs: "WFS",
+    draw: "DRAW",
+    geoJson: "GEOJSON",
+    vectorBase: "VECTORBASE"
+};
+
+export default layertypes;
