+++ exporter/constants/steps.js
@@ -1,0 +1,6 @@
+const steps = {
+    selectLayer: "selectLayer",
+    downloadLayer: "downloadLayer"
+};
+
+export default steps;
