+++ exporter/locales/de/additional.json
@@ -1,0 +1,22 @@
+{
+    "modules": {
+        "tools": {
+            "exporter": {
+                "title": "Export",
+                "prev": "Zurück",
+                "next": "Weiter",
+                "finish": "Abschließen",
+                "drawLayerText": "Importierte Layer",
+                "loadingText": "Lädt…",
+                "layerSelectionText": "Welcher Layer soll exportiert werden?",
+                "emptyLayerSelectionText": "Keine exportierbaren Layer gefunden.",
+                "emptyLayerSelectionTypeText": "Keine Layer vorhanden.",
+                "layerSelectionRequiredText": "Bitte selektieren Sie einen Layer.",
+                "formatSelectionText": "Bitte wählen Sie ein Export-Format aus:",
+                "downloadFailedForFormat": "Export fehlgeschlagen. Bitte probieren Sie ein anderes Format.",
+                "shapeUnsupportedMultiPolygon": "Es können leider keine MultiPolygone im Shape Format exportiert werden. Bitte wählen Sie ein anderes Format.",
+                "vectorBaseLayerText": "Auswahl-Layer"
+            }
+        }
+    }
+}
