+++ exporter/locales/en/additional.json
@@ -1,0 +1,22 @@
+{
+    "modules": {
+        "tools": {
+            "exporter": {
+                "title": "Export",
+                "prev": "Back",
+                "next": "Next",
+                "finish": "Finish",
+                "drawLayerText": "Imported Layers",
+                "loadingText": "Loading…",
+                "layerSelectionText": "Which layer should be exported?",
+                "emptyLayerSelectionText": "No exportable layers found.",
+                "emptyLayerSelectionTypeText": "No layers available.",
+                "layerSelectionRequiredText": "Please select at a layer.",
+                "formatSelectionText": "Please select an export-format:",
+                "downloadFailedForFormat": "Could not generate the export. Please try another format.",
+                "shapeUnsupportedMultiPolygon": "MultiPolygons cannot be exported in Shape format. Please select another format.",
+                "vectorBaseLayerText": "Selection Layers"
+              }
+        }
+    }
+}
